Fixes sub-classed plugins and signals
